Refactor Node interface: change return type of `xxxQps` method from long to double (#564)
- Update test cases (use assertEquals(e, a, delta) instead)
- Also add `totalPass` method in Node interface

+++ sentinel-core/src/main/java/com/alibaba/csp/sentinel/slots/block/degrade/DegradeRule.java
@@ -184,7 +184,7 @@ public class DegradeRule extends AbstractRule {
         } else if (grade == RuleConstant.DEGRADE_GRADE_EXCEPTION_RATIO) {
             double exception = clusterNode.exceptionQps();
             double success = clusterNode.successQps();
-            long total = clusterNode.totalQps();
+            double total = clusterNode.totalQps();
             // if total qps less than RT_MAX_EXCEED_N, pass.
             if (total < RT_MAX_EXCEED_N) {
                 return true;
